+++ src/packet/ethernet_header.c
@@ -38,7 +38,7 @@ ethernet_header_new(void)
 {
     ethernet_header_t *header = (ethernet_header_t *) header_storage_new(&storage);
     
-    LOG_PRINTLN(LOG_HEADER_ETHERNET, LOG_DEBUG, ("Ethernet header new 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_ETHERNET, LOG_DEBUG, ("Ethernet header new 0x%016" PRIxPTR, (uintptr_t) header));
     
     return header;
 }
@@ -48,7 +48,7 @@ ethernet_header_free(header_t *header)
 {
     if (header->next != NULL)   header->next->klass->free(header->next);
     
-    LOG_PRINTLN(LOG_HEADER_ETHERNET, LOG_DEBUG, ("Ethernet header free 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_ETHERNET, LOG_DEBUG, ("Ethernet header free 0x%016" PRIxPTR, (uintptr_t) header));
     
     header_storage_free(header);
 }
