+++ src/packet/ipv4_header.c
@@ -41,7 +41,7 @@ ipv4_header_new(void)
 {
     ipv4_header_t *header = (ipv4_header_t *) header_storage_new(&storage);
     
-    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("IPv4 header new 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("IPv4 header new 0x%016" PRIxPTR, (uintptr_t) header));
     
     return header;
 }
@@ -51,7 +51,7 @@ ipv4_header_free(header_t *header)
 {
     if (header->next != NULL)   header->next->klass->free(header->next);
     
-    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("IPv4 header free 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_IPV4, LOG_DEBUG, ("IPv4 header free 0x%016" PRIxPTR, (uintptr_t) header));
     
     header_storage_free(header);
 }
