+++ src/packet/ptp2_signaling_tlv_header.c
@@ -40,7 +40,7 @@ ptp2_signaling_tlv_header_new(void)
 {
     ptp2_signaling_tlv_header_t *header = (ptp2_signaling_tlv_header_t *) header_storage_new(&storage);
 
-    LOG_PRINTLN(LOG_HEADER_PTP2_SIGNALING_TLV, LOG_DEBUG, ("PTP2 signaling TLV header new 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_PTP2_SIGNALING_TLV, LOG_DEBUG, ("PTP2 signaling TLV header new 0x%016" PRIxPTR, (uintptr_t) header));
 
     return header;
 }
@@ -50,7 +50,7 @@ ptp2_signaling_tlv_header_free(header_t *header)
 {
     if (header->next != NULL)   header->next->klass->free(header->next);
 
-    LOG_PRINTLN(LOG_HEADER_PTP2_SIGNALING_TLV, LOG_DEBUG, ("PTP2 signaling TLV header free 0x%016" PRIxPTR, (unsigned long) header));
+    LOG_PRINTLN(LOG_HEADER_PTP2_SIGNALING_TLV, LOG_DEBUG, ("PTP2 signaling TLV header free 0x%016" PRIxPTR, (uintptr_t) header));
 
     header_storage_free(header);
 }
